pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "approval-factory"
version = "0.1.0"
description = "Approval-first multi-agent development orchestration MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7,<9",
]

[project.scripts]
factory = "orchestrator.cli:main"

[tool.setuptools.packages.find]
include = ["orchestrator*"]
